Moved auth class to new file
