+++ src/components/SingleComment.tsx
@@ -3,14 +3,13 @@ import './SingleComment.scss';
 import Comment from '../Models/Comment';
 import { IonItem, IonAvatar, IonLabel, IonText, IonIcon } from '@ionic/react';
 import { arrowUndo } from 'ionicons/icons';
-import { Timestamp } from '../Models/firebase';
 
 interface CommentProps {
     comment: Comment;
 }
 
 const SingleComment: React.FC<CommentProps> = (props: CommentProps) => {
-    console.log((props.comment.timestamp?.seconds ?? 0) - Timestamp.now().seconds);
+    console.log(props.comment.user.photoUrl);
     return (
         <IonItem className="commentItem">
             <IonAvatar className="commentAvatar">
